TG Shop/requirements.txt
aiogram==3.*
fastapi==0.*
uvicorn[standard]==0.*
httpx==0.*
python-dotenv==1.*
